# perf: process read receipts concurrently